+++ src/app/core/+store/cart/cart.state.ts
@@ -1,0 +1,5 @@
+import {EntityState} from "@ngrx/entity";
+import {CartItem} from "../../../cart/model/cart-item.model";
+
+export interface CartState extends EntityState<CartItem> {
+}